Finished implementing inventory
Finished implementing inventory backend and fixing discovered bugs

+++ Assets/src/Sophia/ItemClass.cs
@@ -27,7 +27,7 @@ public class ItemClass
 
     public virtual Sprite spriteImage{get;}
     public virtual (int str, int len)tempStrength{get;}
-    public virtual (int spd, int len)tempSpeed{get;}
+    public virtual (float spd, int len)tempSpeed{get;}
