+++ Assets/src/Sophia/Slime.cs
@@ -26,7 +26,7 @@ public class Slime : ItemClass
 
     private Sprite subSpriteImage;
     private (int str, int len) subTempStrength;
-    private (int spd, int len) subTempSpeed;
+    private (float spd, int len) subTempSpeed;
 
 /*
 * Summary: Assigns values to item variables, and sets sprite image
@@ -68,7 +68,7 @@ public class Slime : ItemClass
         get {return subTempStrength;}
     }
 
-    public override (int, int) tempSpeed
+    public override (float, int) tempSpeed
     {
         get {return subTempSpeed;}
     }
